move section related code into section folder
