+++ src/packages/media/manifests.ts
@@ -1,5 +1,5 @@
 import { manifests as mediaManifests } from './media/manifests.js';
-import { manifests as mediaSectionManifests } from './section.manifests.js';
+import { manifests as mediaSectionManifests } from './media-section/manifests.js';
 import { manifests as mediaTypesManifests } from './media-types/manifests.js';
 import { manifests as imagingManifests } from './imaging/manifests.js';
 import type { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
