+++ src/packages/media/media-section/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_MEDIA_SECTION_ALIAS = 'Umb.Section.Media';
